Simple video to audio converter
Converting videos to audio files might seem like an odd decision, but it can come in handy in specific cases. It is most often used to record the soundtrack of videos or to extract other audio tracks from videos where you are only interested in the sound. So, with Python, it is super easy as this!